fix(review): address PR review comments for fieldpath, extractor patterns, and documentation
