+++ pkg/common/structured/reader.go
@@ -19,9 +19,7 @@ import (
 	"errors"
 	"fmt"
 	"strconv"
-	"strings"
 	"time"
-	"unique"
 
 	"github.com/GoogleCloudPlatform/khi/pkg/common"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -30,9 +28,6 @@ import (
 
 // ErrFieldNotFound is returned when a requested field is not found in the node structure.
 var ErrFieldNotFound = errors.New("field not found")
-
-// EmptyFieldPath represents an empty field path pointing to the root node itself.
-var EmptyFieldPath = FieldPath{}
 
 // NodeReaderChildrenIterator is a type that represents an iterator function for navigating
 type NodeReaderChildrenIterator = func(func(key NodeChildrenKey, value NodeReader) bool)
@@ -64,32 +59,6 @@ func (n *NodeReader) WithKeyOrder(priorityKeys ...string) *NodeReader {
 	return &NodeReader{
 		Node: WithKeyOrder(n.Node, priorityKeys...),
 	}
-}
-
-// FieldPath represents a pre-parsed and interned field path for fast traversal on StandardMapNode.
-type FieldPath struct {
-	segments []string
-	handles  []unique.Handle[string]
-}
-
-// CompileFieldPath parses and interns a field path string (e.g., "resource.labels.node_name").
-func CompileFieldPath(s string) FieldPath {
-	segments := ParseFieldPath(s)
-	handles := make([]unique.Handle[string], len(segments))
-	for i, seg := range segments {
-		handles[i] = unique.Make(seg)
-	}
-	return FieldPath{segments: segments, handles: handles}
-}
-
-// Segments returns the underlying path segments.
-func (p FieldPath) Segments() []string {
-	return p.segments
-}
-
-// IsEmpty returns true if the field path is empty.
-func (p FieldPath) IsEmpty() bool {
-	return len(p.segments) == 0
 }
 
 // Has checks if a field exists at the pre-compiled FieldPath.
@@ -313,45 +282,6 @@ func ReadReflectK8sRuntimeObject[T runtime.Object](r *NodeReader, path FieldPath
 		return fmt.Errorf("failed to decode JSON as runtime.Object: \n source: %s\nerror:%s", string(rawJSON), err.Error())
 	}
 	return nil
-}
-
-// ParseFieldPath splits a field path string according to specified rules.
-// It uses '.' as a delimiter, but '\.' is treated as an escaped literal dot.
-func ParseFieldPath(s string) []string {
-	var result []string
-	var currentSegment strings.Builder
-	isEscaped := false
-
-	for _, r := range s {
-		if isEscaped {
-			if r == '.' {
-				currentSegment.WriteRune('.') // '\.' is treated as a literal '.' and added to the current segment
-			} else {
-				// If '\' is followed by something other than '.', treat '\' as a literal character too
-				currentSegment.WriteRune('\\')
-				currentSegment.WriteRune(r)
-			}
-			isEscaped = false
-		} else {
-			switch r {
-			case '\\':
-				isEscaped = true
-			case '.':
-				result = append(result, currentSegment.String())
-				currentSegment.Reset() // Reset the current segment
-			default:
-				currentSegment.WriteRune(r)
-			}
-		}
-	}
-
-	if isEscaped {
-		// If the string ends with '\', treat it as a literal '\'
-		currentSegment.WriteRune('\\')
-	}
-	result = append(result, currentSegment.String())
-
-	return result
 }
 
 func getScalarAs[T any](scalarNode Node) (T, error) {
