+++ pkg/common/structured/reader_test.go
@@ -288,72 +288,6 @@ complex:
 	})
 }
 
-func TestParseFieldPath(t *testing.T) {
-	testCases := []struct {
-		name     string
-		input    string
-		expected []string
-	}{
-		{
-			name:     "simple path",
-			input:    "a.b.c",
-			expected: []string{"a", "b", "c"},
-		},
-		{
-			name:     "empty path",
-			input:    "",
-			expected: []string{""},
-		},
-		{
-			name:     "path with single segment",
-			input:    "single",
-			expected: []string{"single"},
-		},
-		{
-			name:     "escaped dots",
-			input:    "a\\.b.c",
-			expected: []string{"a.b", "c"},
-		},
-		{
-			name:     "backslash following non dot char",
-			input:    "a\\_b.c",
-			expected: []string{"a\\_b", "c"},
-		},
-		{
-			name:     "multiple escaped dots",
-			input:    "a\\.b\\.c.d",
-			expected: []string{"a.b.c", "d"},
-		},
-		{
-			name:     "trailing escape character",
-			input:    "a.b\\",
-			expected: []string{"a", "b\\"},
-		},
-		{
-			name:     "trailing escaped dot",
-			input:    "a.b\\.",
-			expected: []string{"a", "b."},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			result := ParseFieldPath(tc.input)
-
-			if len(result) != len(tc.expected) {
-				t.Errorf("Expected %d segments, got %d", len(tc.expected), len(result))
-				return
-			}
-
-			for i, segment := range result {
-				if segment != tc.expected[i] {
-					t.Errorf("Segment %d: expected %q, got %q", i, tc.expected[i], segment)
-				}
-			}
-		})
-	}
-}
-
 func TestFieldPath(t *testing.T) {
 	yamlData := `
 string_value: "test string"
